moving folders inside of keyvalue directory
