+++ temporal/keyvalue/internal/driver/redisv9/redisv9.go
@@ -1,8 +1,8 @@
 package redisv9
 
 import (
-	"github.com/TykTechnologies/storage/temporal/internal/driver/rediscommon"
-	"github.com/TykTechnologies/storage/temporal/internal/types"
+	"github.com/TykTechnologies/storage/temporal/keyvalue/internal/driver/rediscommon"
+	"github.com/TykTechnologies/storage/temporal/keyvalue/internal/types"
 	"github.com/redis/go-redis/v9"
 )
