+++ temporal/keyvalue/internal/driver/redisv9/redisv9_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/TykTechnologies/storage/temporal/internal/types"
+	"github.com/TykTechnologies/storage/temporal/keyvalue/internal/types"
 )
 
 func TestNewRedisV9(t *testing.T) {
